+++ packages/tswebextension/src/lib/mv2/background/web-request-api.ts
@@ -275,6 +275,7 @@ export class WebRequestApi {
             frameId,
             requestUrl,
             referrerUrl,
+            eventId,
             requestId,
             contentType,
             timestamp,
@@ -291,7 +292,7 @@ export class WebRequestApi {
             type: FilteringEventType.SendRequest,
             data: {
                 tabId,
-                eventId: requestId,
+                eventId,
                 requestUrl,
                 requestDomain: getDomain(requestUrl),
                 frameUrl: referrerUrl,
@@ -339,7 +340,7 @@ export class WebRequestApi {
         // the response in order to actually apply $replace rules to it.
         const response = RequestBlockingApi.getBlockingResponse(
             basicResult,
-            requestId,
+            eventId,
             requestUrl,
             requestType,
             tabId,
@@ -439,11 +440,15 @@ export class WebRequestApi {
         context,
         details,
     }: RequestData<WebRequest.OnHeadersReceivedDetailsType>): WebRequestEventResponse {
+        if (!context) {
+            return undefined;
+        }
+
         defaultFilteringLog.publishEvent({
             type: FilteringEventType.ReceiveResponse,
             data: {
-                tabId: details.tabId,
-                eventId: details.requestId,
+                tabId: context.tabId,
+                eventId: context.eventId,
                 statusCode: details.statusCode,
             },
         });